ch5: Shader::set_uniform expects U to be const-ref
Introduced the input types glm::vec2-4 and glm::mat2-4 for
Shader::set_uniform. These types are larger and make sense to reference
instead of copy when setting them to uniforms in a shader program
however, this introduced ambiguity when calling the method due to
expected different cv-qualifiers for scalar inputs and vector and matrix
inputs. Opted to have all inputs required to be const-ref as it wouldn't
introduce any cost or reduce usability for scalar inputs as const-ref
still permits literals.